Lots of delinting
With the help of clang-tidy this patch mostly is cleaning up coding
style problems and making a few recommended changes.

Using is prefered to typedef in almost all situations. I have also added
in copy/move constructors that were missing (mostly deleted ones). I
have also changed the include arrangement to the way LLVM wanted them.

Boost Range is not being used anymore after it was discovered it wasn't
actually filtering the way it was supposed to. The `locked` container
functions now return a slist designed to be iterated over in the forward
direction.

There also now some new helper functions for qDebug to accept
std::string and grpc::string_ref as clang-tidy hates printf style format
strings. It also hates numeric literals in the code; almost all of them
have been replaced with named constants or, in the case of times, using
tools from std::chrono to change durations.

Now that all the dirt has been removed I think my next direction will be
to look at the API itself. It's still using protobuf v2, while gRPC
itself is designed for v3. Also missing are ways to gracefully
disconnect from a streaming interface. Good for finding bugs but not
good from a user-perspective.

If any reviewers are not liking the direction this patch set is going,
please let me know now; I have tried to keep the external changes
to a bare minimum so far but I believe the needs changes that would
break existing clients to get this to where it would be a better and
more usable interface than dbus or ICE.