PulseAudio: cork input when it is not required
When self-muted or deafened, "cork" the audio input. Corking
instructs the audio source to stop producing audio samples which—
when the user is muted—are not needed and would never be
transmitted.

Without samples to process, the input DSP doesn't run. Corking can
halve mumble's CPU consumption when muted, but your mileage may
vary.

This patch adds the MainWindow::corkAudioInputStream() signal to
request changes in the corking state. Other audio backends could
be modified to connect to this signal.