mumble/VersionCheck: Made assembling of RequestURL run in separate thread
As it stands, the VersionCheck always causes small mini-freezes. The
part that is responsible for them is reading the application binary
and computing its hash-value which has to be attached to the
request-URL.
Thus this whole process is outsourced to another thread so to not
delay Mumble's startup time.

Thanks to Qt's singal/slot mechanism, the processing of the incoming
answer to the request is handled in the main-thread (just as before).