Channel Listeners: Added local volume adjustment
With this commit the possibility to perform local volume adjustments to
different listener proxies is introduced. This functionality was
previously only possible for other clients on the server.

You can open the UI element for adjusting the volume via the context
menu of your listener proxy.

These volume adjustments will be applied to all audio streams that are
being received via the listener proxy (aka the ones you only hear
because you are listening to that channel).