src/mumble/ACLEditor.cpp: Always modify Listen if Enter has been modified
We don't want to end up in situations in which an admin denies someone
Enter privilege into a channel but forgets to also deny Listen privilege
leading to the user still being able to eavesdrop into conversations in
that channel.
Thus if one explicitly allows or denies the Enter privilege, Listen will
be set to the same value automatically.

If the admin explicitly wants the Listen privilege to be different from
the Enter one, (s)he can still do so by manually setting the Listen
privilege after having modified Enter.