src/mumble/UserModel.cpp: Fix talking indicator
Right now Mumble can come in a state where the local client is muted but
its talking state is still frozen to Talking.
This is also reflected in the UI (#4006) so this commit provides a
workaround that at least fixes the UI symptoms of the bug as I haven't
been able to track down the actual origin of it.