src/mumble/WebFetch.cpp: Fix endless loop if there is no prefix
If we don't have a "servicePrefix" set, prefixedServiceHost() and
serviceHost() return the same host. The code however expected them to
implicitly be different.
This only ever made a difference if the main service host wasn't
reachable but in that case the endless-loop kicks in and the CPU usage
goes up.