src/mumble/Log.cpp: Fix message spacing after frame
Right now the first message logged after a message that has been logged
with a frame drawn around it, has a comparatively huge spacing to said
frame.
This is because in the old implementation an additional empty block
sneaked in after the frame, causing the spacing.
With this commit it is checked whether there already exists an empty
block that can be written into before creating a new one.