src/mumble: Added & implemented setting for message margin
This commit adds a setting for specifying the margins for logged
messages. With the default value for this, messages are also spaced out
more than they have been before this commit (if you don't like it, set
the margin to 0 and you're back at the old style).

This commit also changes the framestyle to dashed in order to make
frames less intrusive on the eye if they have to be drawn.