src/mumble/ClientUser.cpp: Narrow down the scope of QWriteLocker
In ClientUser::remove a QWriteLocker is created on
ClientUser::c_qrwlUsers in order to remove the respective user from
ClientUser::c_qmUsers.
However the lock is not released after that but it is instead held until
the function exits. This causes the lock to be held even when
AudioOutput::removeBuffer is being called. Apart from this being bad
practice, it can also lead to a deadlock (verified for the case in which
a ReadLock on ClientUser::c_qrwlUsers is being requested inside
AudioOutput::mix [after the lock on AudioOutput::qrwlOutputs has been
acquired]. This can happen if one calls e.g. ClientUser::get in that
function. In that case the deadlock can happen if audio is being
received and played while one or more users are disconnecting from the
server).