src/mumble/Overlay*: Don't hook into any process before the Overlay is active
Right now the Overlay will start hooking into processes on its
construction. This will cause this to happen even if the overlay is
actually disabled. This is a rather intrusive behavior and can lead to
warnings of the OS (for now macOS only - #3552).

This commit refactors the Overlay class so that it will only start
hooking into processes, if it is actually enabled (that is setActive is
called with true).

This commit is an adapted version of the commit merged into master in so
far as it doesn't use std::atomic which is a cpp11 feature and we don't
require cpp11 for the 1.3.x series.

src/mumble/