scripts/azure-pipelines/build_macos.bash: Static linkage for Mumble on OSX
The published installer is more or less useless as it contains a version
of the Mumble executable that is dynamically linked meaning that it
requires the dependencies to be installed on the Host it's being
installed on.

This commit changes the CI build for MacOS to use static linkage in
order to allow the installer to be used normally.