src/mumble: Implemented #1548: Whisper/Shout to selection
Before it was only possible to set hotkeys for whispering/shouting to
fixed targets (e.g. a certain list of users or a certain channel).

This commit adds the possibility to set up a hotkey to whisper/shout to
whatever is the current selection in the UI. That means if I have
another user selected, then I'll whisper to that user and if I have a
channel selected, then I'll shout to that instead.