src/mumble/AudioInput.cpp: Fix whisper-target logic interfering with normal speech
In order to send the last audio frame of a whisper-action to the correct
target, we have to do some workaround introduced in
9d16137393bd8a33307cac41c0b55ea1392173a3. The current implementation
bleeds into normal speech, if VoiceActivation is used (#2604).

This is due to iPrevTarget not getting reset automatically once it has
fulfilled its purpose. If PTT is used as transmission mode, it doesn't
matter as the button-pushing also resets iPrevTarget as needed.
VoiceActivation doesn't do that though, so this commit adds a little
housekeeping to the iPrevTarget logic.