src/mumble/AudioInput: Fix racecondition for g.iTarget
g.iTarget is not protected by a lock or a Mutex. This means that it can
change its value while AudioInput::encodeAudioFrame is running. However
the value is accessed at multiple stages (including
AudioInput::flushCheck that is called from encodeAudioFrame) meaning
that there could be inconsistencies of what g.iTarget was meaning for
this audio frame.

This commit makes sure that g.iTarget is only accessed once at the
beginnign of AudioInput::encodeAudioFrame and then copied to a local
variable which is then used for all the places in which g.iTarget
previously was access directly.

One scenario this race condition could be observed was pressing and
immediately releasing a whisper shortcut. This could sometimes cause one
of the whisper frames (the second last to be more precise) to be sent as
normal audio instead of whispering (the last message would have been
affected by this as well however for that there's a special workaround
in place already [9d16137393bd8a33307cac41c0b55ea1392173a3] which
covered the effects of the racecondition for that one).