src/mumble/AudioOutputSpeech.cpp: Set talking state to passive in destructor
Right now if an AudioOutputSpeech object gets deleted while not having
finished playing audio, the talking state of the corresponding user will
freeze to be Talking when in fact it should be Passive.

This commit sets the talking state of the corresponding user to Passive
in the destructor to avoid this issue.