src/mumble: Implemented an optional TalkingUI
The TalkingUI is meant as some sort of a light version of the Overlay
primarily intended for use when using the computer normally (e.g. with
many different programs instead of one that is FullScreen).

It consists of a small window that can be toggled and it will show the
names of the person currently talking as well as the channel that person
is in.

If a person has stopped talking for some time (5s by default but this
interval is configurable via the settings), it will be removed from
the UI again.

By default the UI will stay on top of all other windows so that you'll
always be able to see it.

It is possible to select a user in the TalkingUI. This will lead to the
user being selected in the main ChannelTree as well. The other way
around works as well. If however the user selects a channel in the main
ChannelTree, nothing will be selected in the TalkingUI.
This selection makes it possible to use the TalkingUI in combination
with the "Whisper to selection" feature.