src/murmur/MurmurGRPCImpl.h: Use namespace grpc instead of grpc_impl
It appears as if more modern versions of GRPC-c++ have moved their
classes from namespace grpc to namespace grpc_impl, which is what the
most recent contribution to this file was using. However in older
versions of the GRPC library this namespace doesn't exist yet and thus
causes compile errors.

Luckily it seems however that the GRPC devs were smart about this
change as they added typedefs to the grpc namespace (which still exists)
that reroute to the grpc_impl namespace. Thus we can simply use the grpc
namespace and in newer versions this will redirect us to the grpc_impl
namespace automatically.

As a reference on the GRPC side see
https://github.com/grpc/grpc/commit/d684ddc7e3091b8b5205b0cadbd786bd5f37487d