src/mumble: Added setting for TalkingUI to keep local user visible
With this setting enabled, the local user is always visible in the
TalkingUI as long as the user is connected to a server.

Furthermore the local user will always be listed first in its channel
and the channel of the local user will also be at the top of the
TalkingUI. This is independent from the new setting.

In order to make sure that the state displayed in the TalkingUI is
actually correct, some EventHandlers are added to the TalkingUI as well.