src/mumble: Don't hide the setting if no-update config was used
With commit 847ed7ad0eef3c4e1a4a7dd230e05b70f17ec412 it was introduced
that when using the no-update config flag, not only the default values
for checking or not checking for updates were affected but actually the
whole setting would be hidden from the UI.

This commit reverts that as a user should always be able to decide for
him- or herself whether or not they want to see update notifications or
not.

It does make sense for package maintainers to disable these checks by
default (as updates are received via package manager anyways) but if the
user wants to have that check enabled nonetheless, they should be able to
do so.