src/mumble: Implemented options to display channel hierarchy in TalkingUI
Instead of only showing the name of the channel the respective user is
currently in, this commit adds the possibility to also include (parts
of) the channel hierarchy to remove ambiguities.

Alongside comes a system to potentially abbreviate the (hierarchy) names
in order to not use too much screen space.