src/mumble/TalkingUI.cpp: Allow user-context menu to be triggered from the TalkingUI
This commit uses a "hack" to trick the MainWindow to display the context
menu for a given user in the TalkingUI. That's very convenient for
people that have only the TalkingUI open and want to take quick actions
(e.g. adjust volume of a user, mute a user, etc.) without having to open
the MainWindow again.