src/mumble/MainWindow: Added context action to join user's channel
This might be useful for situations in which a channel is pretty full
and thus one doesn't see the channel entry in the UI but instead sees
some users that are in that channel. Now if you want to join these
users, you don't really care about which channel they are in anyway, so
with this you can just right click on any user in a different channel
and join their channel via the context menu.