src/mumble/UserModel: Fixed rowOf() function for listener proxies
The ModelIndex::rowOf functions did not take into account that checking
whether ModelIndex::pUser is not sufficient anymore as listener proxies
share the same associated user. Thus to uniquely identify the index for
the actual user, it also has to be taken into account whether the index
at hand corresponds to a listener proxy.

This mis-behavior was responsible for a segfault triggered by being in a
channel you also have a listener proxy in and then being dragged out of
that channel (the dragging part is important). During this process it
would delete the wrong item which it later wanted to select again
(causing access to deleted memory) resulting in the segfault.