src/mumble/TalkingUI.cpp: Add space between channel name and first entry
Before there was no space at all between the channel name and the first
entry in the list. This was especially apparent when selecting that
first entry. This could cut of parts of the channel name (especially
when using very small font sizes).

This commit adds some spacing in order to make things look less clamped.