src/mumble/AudioOutput.cpp: Do not assume Y up for small errors
If front and top are not perpendicular by a small margin, it is better
to find a perpendicular vector that differs the least from top.

Only for large errors, i.e. if front and top are not perpendicular at
all, it makes sense to create a totally new top vector by assuming Y is
up.

This fixes distorted audio if the positional audio plugin would
sometimes (but only rarely) deliver errors larger than 0.1, in which
case Mumble would rapidly switch between Y up and the real up vector.

Fixes #4169