src/mumble/NetworkConfig: Show warning message for enabling update checks with no-update config
If Mumble has been compiled with the no-update config flag, then the
auto-updates are deactivated by default. If a user wants to enable it in
the settings, a pop-up will be shown to the user explaining that the
update notifications are probably not meaningful (as the user is most
likely getting their Mumble updates from a package repository). This is
to prevent confusion of the user.