src/mumble/Talking.cpp: Fix race leading to crash
The problem that can arise is the following: A user changes channel.
Inside Messages.cpp UserModel::moveUser gets called that directly
updates the channel of the user. In order to propagate this change to
the TalkingUI, QMetaObject::invokeMethod is used leading to an
asynchronous processing of this event. If however in the same iteration
of the event loop the audio system sent a change in TalkingState, that
will be processed first in the next iteration of the event loop (before
the channel change is processed). This violates the assumption of
TalkingUI::ensureVisible that the channel of the local user is always
present in the TalkingUI. Due to how HashMaps work for non-existent keys
when using the [] operator, a nullptr gets inserted in the HashMap. If
that gets dereferenced at a different point in the code, we get a
SegFault.

This commit prevents this by first explicitly adding the local user's
channel first. This covers the edge case described above.