src/mumble/TalkingUI.cpp: Made code more robust against non-existent hash entries
The [] operator of QHash was mostly removed from the code and in those
places where it is still used, an explicit check of whether the entry in
question really exists is done beforehand.

Thus instead of inserting weird entries into the hash (if a given key
doesn't have a mapping yet), an error message is printed to the console
log and the respective functions exits.

Overall this should make the code more robust (crash-wise) against
non-existing hash entries as these should now lead to an error message
(and maybe weird states in the TalkingUI) instead of an application
crash.