src/murmur/ServerDB: Make sure the Database uses proper encoding
In #4213 it was reported that connecting to a server with a unicode
character (Emoji) in the name can crash the server (provided the user
name passes validation which is only possible if the server uses a
special rule for the username config) when using MySQL as the backend.

This is because we have been using "utf8" for MySQL which - as it turns
out - isn't actual utf8 (https://mathiasbynens.be/notes/mysql-utf8mb4).

Therefore this commit changes this behavior by setting the proper utf8
encoding for MySQL databases. For SQLite and PostgreSQL the encoding is
checked and if it is not utf8 (or utf16), the server will refuse to
start. This is because there is no way to set the encoding for existing
databases for these backends.