src/murmur: Added config option autostart and bootserverid
With these options it can be controlled whether all available virtual
servers shall be booted automatically (autostart) or if only a specific
server should be booted (bootserverid).

If bootserverid is not set and autostart is set to false, no server is
booted. The user then has to explicitly boot a server using e.g. ICE or
gRPC.