ACL:Add new permission for Reset Comment/Avatar in server & client
This commit introduces a new permission, ResetUserContent, to grant the ability to reset another user's comment or avatar.
It also includes a check in the ACL editor window. If the server version is less than 1.4.0 then don't show the new permission in the ACL checkbox list as the server won't respond to it.
On older servers the Move permission in the root channel was also used for allowing or denying comment and avatar changes.