src/mumble/TalkingUI.cpp: Fix "empty entry bug"
The bug appeared occasionally when changing channel into another one in
which there was someone talking. In that case it could happen that in
the TalkingUI the old channel the local user has just moved out from,
would still show up and remain in the TalkingUI (until another action -
change of talking state or change of channel) even though that channel
was empty now.

This was due to a race condition between the channelChanged and the
talkingStateChanged event. If the talkingStateChanged event arrived
first, the code that is responsible for inserting the local user at the
top of the channel it is in, would then blindly "remove" the local user
from that channel in order to add it again at the top. In this situation
however, the local user wasn't even in that channel yet (in the
TalkingUI that is) so the "remove" actually didn't do anything whereas
the adding now added the user to a channel it wasn't in before. As this
bypassed all mechanisms for normal channel change (that include
potential clean-up of old, empty channels), the old channel would not be
hidden by this.

This commit now simply checks whether the local user is actually in the
channel in question already (in the TalkingUI) and only if it isn't
this part of the code is simply skipped assuming that the channelChanged
event for the local user will be processed later causing it to land at
the top of that channel eventually.

Before this commit, the bug could be most easily reproduced if one
commented out the body of TalkingUI::on_channelChanged. Then you need 2
clients (transmit mode = continuous) on the same server. Make sure you
start out in different channels. Then change channel of one client to
the channel of the other client. The bug most often occurred on the
first try. If not, try again (use mute/unmute to update the channel of
the local user in the TalkingUI in between).