src/mumble/AudioOutput.ui: Move positional audio option to proper section
We do have a separate section in the AudioOutput settings for positional
audio but the option to enable/disable it is not in that section. This
doesn't make a whole lot of sense.

This commit simply moves the setting into the positional audio section
and also renames it from "positional audio" to "enable" as due to its
location in the settings, it's clear what it'll be about.