REFAC(crypto): remove QSslSocket::addDefaultCaCertificates
The function in question has been deprecated in Qt 5.15. Instead it is
recommended to set the CA certificates on the QSslConfiguration object.
According to the docs of QSslConfiguration however the system's CA
certificates is what is used by default anyways.

Quote (from 5.15 docs of QSslConfiguration::systemCaCertificates):
"This function provides the CA certificate database provided by the
operating system. The CA certificate database returned by this function
is used to initialize the database returned by caCertificates() on the
default QSslConfiguration."