FIX(overlay): Make overlay work again
In aac3214d4be2ba31c852619bc8d3666391dcd02a Overlay::createPipe() has
been introduced in order to only create the pipe when needed (instead of
in the constructor of the Overlay class). This new function was never
called though which caused the Overlay to stop working.

This commit makes sure the respective function is called when activating
the overlay.

Fixes #4281