FIX(certificate): Properly process signed certificates
This fixes a regression introduced in
b691108d3820695d51e2f4d76d1ee080e9bfaa91. In there it was thought that
there is no need to explicitly add the system CaCertificates to the SSL
configuration. This turned out to be wrong.

It was brought up that by removing this line, properly signed
certificates would also be shown as self-signed.

Fixes #4289