FEAT(protocol): Add Capabilities message to conduct voice protocol negotiation.
See Mumble.proto for more details. Mumble doesn't support other voice protocol so far,
so this message is intended to be used to negotiate the cipher type used in transportation.
It is sent between Version and Authenticate message, before the CryptState is sent from
the server when handling the Authenticate message.
A config option "voiceProtocolPreference" is also added to allow the server administrator
to specify the protocol allowed by the server.

This feature implements the idea of #3918 and #4248.