FEAT(serverbrowser): Update serverbrowser UI
The location subtree for the public servers is now gone.
All public servers are now child elements of "Public Internet" regardless
of location. A flag Icon has been added to each public server to
indicate its location.

A search bar for the public servers is added.  It is only visible when the
public servers are expanded.  Searching for a server name and filtering
by location is possible. The "server status filters"
(Show Reachable etc...) have been moved from the context menu of the
server browser to the search dialog.

This also adds a window that is shown once the user expands the public
server list for the first time. The user is asked for consent to
transmit the IP address to each of the public servers.
If the user does not consent, the public server list is disabled.

Implements #3532