FIX(talking-ui): Fix glitch on session reuse
Until now the TalkingUI assumed that session IDs of users are unique
throughout the server's lifetime. This turned out to be wrong as the
server reuses session IDs of disconnected users.

This could lead to a new user being shown with the name of an old user
in the TalkingUI as that new user was reusing the old's session ID.

With this commit a user gets deleted from the TalkingUI if it
disconnects from the server.