FIX(talking-ui): Fix local user added multiple times
In #4306 a glitch in the TalkingUI was fixed that was associated with
session reuse by the server. As an additional fail-safe the respective
commit (bbfe736b133201c685ecc86dbbec5079aa458c18) also implemented a
quick check for whether the name of the user with a given ID is still
the same as the stored one for that user.

In #4325 however it was pointed out that this leads to the local user
being added to the TalkingUI every time it starts speaking. This is due
to the local user's name being written in bold meaning that it contains
HTML tag in the TalkingUI which are obviously not part of the actual
user name.

This commit fixes this issue by first stripping out any HTML tags from
the stored name before comparing it to the actual username.

Fixes #4325